requirements.txt
apache-beam[GCP]
google-cloud-resource-manager==0.30.3
google-cloud-storage==1.38.0
google-cloud-logging==3.0.0
bigquery-schema-generator==1.4
